requirements.txt
keras==2.10.0
numpy==1.23.3
pandas==1.0.5
Pillow==9.3.0
streamlit==1.15.2
streamlit_drawable_canvas==0.9.2
tensorflow==2.10.0